+++ iOS/G3MiOSSDK/Commons/Basic/MultiLayerTileTexturizer.cpp
@@ -399,7 +399,7 @@ public:
       std::vector<const IImage*>     images;
       std::vector<RectangleF*> sourceRects;
       std::vector<RectangleF*> destRects;
-      std::vector<double> transparencies;
+      std::vector<float> transparencies;
       std::string textureId = _tile->getKey().tinyDescription();
 
       const Sector tileSector = _tile->_sector;
