+++ iOS/G3MiOSSDK/Commons/Layers/MercatorTiledLayer.hpp
@@ -45,7 +45,7 @@ public:
                      int                             initialLevel,
                      int                             maxLevel,
                      LayerCondition*                 condition,
-                     double transparency = 1.0);
+                     float transparency = 1.0);
 
   URL getFeatureInfoURL(const Geodetic2D& position,
                         const Sector& sector) const;
