+++ iOS/G3MiOSSDK/Commons/Layers/TMSLayer.cpp
@@ -26,7 +26,7 @@ TMSLayer::TMSLayer(const std::string& mapLayer,
                    const TimeInterval& timeToCache,
                    bool readExpired,
                    const LayerTilesRenderParameters* parameters,
-                   double transparency):
+                   float transparency):
 
 Layer(condition,
       mapLayer,
