+++ iOS/G3MiOSSDK/Commons/Layers/TMSLayer.hpp
@@ -38,7 +38,7 @@ public:
            const TimeInterval& timeToCache,
            bool readExpired,
            const LayerTilesRenderParameters* parameters = NULL,
-           double transparency = 1.0);
+           float transparency = 1.0);
 
   std::vector<Petition*> createTileMapPetitions(const G3MRenderContext* rc,
                                                 const LayerTilesRenderParameters* layerTilesRenderParameters,
