+++ iOS/G3MiOSSDK/Specific/Derived/Canvas_iOS.hpp
@@ -106,7 +106,7 @@ protected:
   void _drawImage(const IImage* image,
                   float srcLeft, float srcTop, float srcWidth, float srcHeight,
                   float destLeft, float destTop, float destWidth, float destHeight,
-                  double transparency);
+                  float transparency);
 
 
   void _beginPath();
